[ATOM-vLLM] Upgrade vLLM version to v0.22.0 (#1006)
upgrade atom-vllm vllm version to 0.22.0
