+++ .github/workflows/docker-release.yaml
@@ -93,8 +93,8 @@ jobs:
       RCCL_REPO: "https://github.com/ROCm/rccl.git"
       RCCL_BRANCH: "29e1567b95e28823b0beb1a988adc587bfab5b4f"
       GPU_ARCH: "gfx942;gfx950"
-      VLLM_COMMIT: "2a69949bdadf0e8942b7a1619b229cb475beef20"
-      VLLM_VERSION: "0.19.0"
+      VLLM_COMMIT: "0b3ba88f165976e77ca5e6a7a3f5bba4562b80af"
+      VLLM_VERSION: "0.22.0"
       SGLANG_REPO: "https://github.com/sgl-project/sglang.git"
       SGLANG_REF: "v0.5.10"
       SGLANG_VERSION: "0.5.10"
